+++ test/fonts/test_fonts.py
@@ -45,6 +45,17 @@ class FontsTest(unittest.TestCase):
                 pdf.text(0, 10 + 40 * i + 10 * j, "Hello World!")
         assert_pdf_equal(self, pdf, "test_set_builtin_font.pdf")
 
+    def test_issue_66(self):
+        pdf = FPDF()
+        pdf.add_page()
+        pdf.set_font("Times", "B", 14)
+        pdf.cell(50, 0, "ABC")
+        pdf.set_font("Times", size=10)
+        pdf.cell(50, 0, "DEF")
+        # Setting the font to an already used one used to remove the text!
+        pdf.set_font("Times", "B", 14)
+        assert_pdf_equal(self, pdf, "test_issue_66.pdf")
+
 
 if __name__ == "__main__":
     unittest.main()